requirements.txt
whisper
openai
ffmpeg-python
python-dotenv
pysubs2